Created Attribute trait and set Entity meta to use
