+++ src/Actinoids/Modlr/RestOdm/Metadata/EntityMetadata.php
@@ -15,6 +15,8 @@ use Actinoids\Modlr\RestOdm\Metadata\Interfaces\PersistenceInterface;
  */
 class EntityMetadata implements AttributeInterface, MergeableInterface
 {
+    use Traits\AttributeTrait;
+
     /**
      * The id key name and type.
      */
@@ -61,14 +63,6 @@ class EntityMetadata implements AttributeInterface, MergeableInterface
      * @param PersistenceInterface
      */
     public $persistence;
-
-    /**
-     * All attribute fields assigned to this entity.
-     * An attribute is a "standard" field, such as a string, integer, array, etc.
-     *
-     * @var AttributeMetadata[]
-     */
-    public $attributes = [];
 
     /**
      * All relationship fields assigned to this entity.
@@ -223,69 +217,6 @@ class EntityMetadata implements AttributeInterface, MergeableInterface
     public function getParentEntityType()
     {
         return $this->extends;
-    }
-
-    /**
-     * Adds an attribute field to this entity.
-     *
-     * @param   AttributeMetadata   $attribute
-     * @return  self
-     * @throws  MetadataException   If the attribute key already exists as a relationship.
-     */
-    public function addAttribute(AttributeMetadata $attribute)
-    {
-        if (isset($this->relationships[$attribute->getKey()])) {
-            throw MetadataException::fieldKeyInUse('attribute', 'relationship', $attribute->getKey(), $this->type);
-        }
-        $this->attributes[$attribute->getKey()] = $attribute;
-        ksort($this->attributes);
-        return $this;
-    }
-
-    /**
-     * Gets all attribute fields for this entity.
-     *
-     * @return  AttributeMetadata[]
-     */
-    public function getAttributes()
-    {
-        return $this->attributes;
-    }
-
-    /**
-     * Determines any attribute fields exist on this entity.
-     *
-     * @return  bool
-     */
-    public function hasAttributes()
-    {
-        return !empty($this->attributes);
-    }
-
-    /**
-     * Determines if an attribute field exists on this entity.
-     *
-     * @param   string  $key
-     * @return  bool
-     */
-    public function hasAttribute($key)
-    {
-        return null !== $this->getAttribute($key);
-    }
-
-    /**
-     * Gets an attribute field from this entity.
-     * Returns null if the attribute does not exist.
-     *
-     * @param   string  $key
-     * @return  AttributeMetadata|null
-     */
-    public function getAttribute($key)
-    {
-        if (!isset($this->attributes[$key])) {
-            return null;
-        }
-        return $this->attributes[$key];
     }
 
     /**
